rename pkg depencencies to deps
